requirements.txt
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.3.0
asyncer==0.0.8
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
diskcache==5.6.3
distro==1.9.0
docker==7.0.0
fast-depends==2.4.12
FLAML==2.1.2
Flask==3.0.3
Flask-Cors==4.0.1
Flask-RESTful==0.3.10
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
markdownify==0.12.1
MarkupSafe==2.1.5
mysql-connector-python==9.2.0
numpy==1.26.4
openai==1.58.0
packaging==24.0
pillow==10.3.0
pyautogen==0.7.2
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
pytz==2024.1
regex==2024.5.10
requests==2.31.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
termcolor==2.4.0
tiktoken==0.6.0
tqdm==4.66.4
typing_extensions==4.11.0
urllib3==2.2.1
websockets==14.2
Werkzeug==3.0.3
